requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
Pillow>=9.0.0
PyMuPDF>=1.20.0
opencv-python>=4.5.0
pdf2image>=1.16.0
pyyaml>=6.0
loguru>=0.6.0
tqdm>=4.64.0
requests>=2.28.0
aiohttp>=3.8.0
fastapi>=0.95.0  # Optional for API server
uvicorn>=0.22.0  # Optional for API server
python-multipart>=0.0.6  # Optional for API server
